Update windows-server-2016-improvements.md
Via lab testing, as well as verifying the code directly, I've confirmed that the W32time service does NOT update configuration related for UtilizeSslTimeData until the service is restarted. This can be seen with Time-Service event 263. 

This article has misleading text regarding this. 

I am proposing to replace this statement: 

"When the setting is picked up by a Group Policy Client, W32time service is notified and it will stop time monitoring and enforcement using SSL time data. The SSL time data collection will stop when each machine reboots." 


With this statement: 

For the time service to update this change and stop SSL time data collection, the client machine requires either a reboot or a w32time service restart.